fix: resolve all 31 clippy warnings
- Gate encode-only imports/functions with #[cfg(feature = "encode")]
- Remove unnecessary u32/i16 casts
- Use .div_ceil(2) instead of (x + 1) / 2
- Allow too_many_arguments in YUV conversion modules (plane/stride
  params are inherent to the domain)
- Allow collapsible_if where #[cfg] prevents collapsing
- Use iter_mut().enumerate() instead of range indexing

+++ src/decoder_managed.rs
@@ -283,8 +283,8 @@ impl ManagedAvifDecoder {
                 .primary_metadata()
                 .map_err(|e| at(Error::from(e)))?;
             (
-                meta.max_frame_width.get() as u32,
-                meta.max_frame_height.get() as u32,
+                meta.max_frame_width.get(),
+                meta.max_frame_height.get(),
             )
         };
